requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.31.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gradio==4.38.1
gradio_client==1.1.0
greenlet==3.0.3
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.5
idna==3.7
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyboard==0.13.5
kiwisolver==1.4.5
langchain==0.2.8
langchain-core==0.2.20
langchain-groq==0.1.6
langchain-text-splitters==0.2.2
langsmith==0.1.87
llm==0.14
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
openai==1.35.14
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
playsound==1.3.0
pluggy==1.5.0
pvporcupine==3.0.2
PyAudio==0.2.14
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pygame==2.6.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-ulid==2.7.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
ruff==0.5.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.36
starlette==0.37.2
sympy==1.13.0
tabulate==0.9.0
tbb==2021.13.0
tenacity==8.5.0
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
websockets==11.0.3
Werkzeug==3.0.3
yarl==1.9.4